Error loc (pydantic/pydantic-core#507)
* Error loc using alias path

* support negative indexes on PathItems

* fix error location on typed-dicts

* cover the required case, add dataclass support

* support arguments

* rename to loc_by_alias

Original-commit-link: https://github.com/pydantic/pydantic-core/commit/7182275dca439c4664c1781758aa533b98c29870

+++ src/input/input_python.rs
@@ -79,7 +79,7 @@ impl<'a> Input<'a> for PyAny {
     fn as_loc_item(&self) -> LocItem {
         if let Ok(py_str) = self.downcast::<PyString>() {
             py_str.to_string_lossy().as_ref().into()
-        } else if let Ok(key_int) = self.extract::<usize>() {
+        } else if let Ok(key_int) = self.extract::<i64>() {
             key_int.into()
         } else {
             safe_repr(self).to_string().into()
